requirements.txt
# ═══════════════════════════════════════════════════════════════════
# ARGUS — Autonomous Real-time Greenhouse Gas Unified Surveillance
# Production requirements — MIT-WPU Ignisia 2024
# ═══════════════════════════════════════════════════════════════════

# ── Core scientific stack ──────────────────────────────────────────
numpy==1.26.4
pandas==2.2.2
scipy==1.13.0
xarray==2024.3.0
netCDF4==1.7.1
h5py==3.11.0

# ── Geospatial ────────────────────────────────────────────────────
geopandas==0.14.4
shapely==2.0.4
pyproj==3.6.1
rasterio==1.3.10
fiona==1.9.6

# ── Google Earth Engine ───────────────────────────────────────────
earthengine-api==0.1.390
geemap==0.31.0

# ── TorchGeo — geospatial deep learning ──────────────────────────
torchgeo==0.5.2

# ── Deep learning ─────────────────────────────────────────────────
torch==2.3.0
torchvision==0.18.0
timm==1.0.3
einops==0.8.0

# ── PyTorch Geometric (GNN) ───────────────────────────────────────
torch-geometric==2.5.3
torch-scatter==2.1.2
torch-sparse==0.6.18

# ── NVIDIA Modulus (Physics-Informed ML) ─────────────────────────
# Install separately: pip install nvidia-modulus
# Listed here for documentation — see Dockerfile for install order
nvidia-modulus==0.6.0

# ── LLM agent (Groq — free tier) ─────────────────────────────────
openai==1.30.0
groq==0.9.0
jinja2==3.1.4

# ── API layer ─────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
httpx==0.27.0

# ── Database ──────────────────────────────────────────────────────
motor==3.4.0
pymongo==4.7.1

# ── Dashboard ─────────────────────────────────────────────────────
streamlit==1.35.0
plotly==5.22.0
keplergl==0.3.2
streamlit-keplergl==0.3.0
folium==0.16.0
streamlit-folium==0.20.0

# ── Utilities ─────────────────────────────────────────────────────
pyyaml==6.0.1
python-dotenv==1.0.1
loguru==0.7.2
tqdm==4.66.4
tenacity==8.3.0
matplotlib==3.9.0
Pillow==10.3.0